Desktop/dicom-viewer/requirements.txt
tensorflow
numpy
pydicom
PyQt5
matplotlib
Pillow